CMPLRLLVM-40254: Fix exposed seg fault in many loops specialization heuristic in IP Cloning. (#7254)
* CMPLRLLVM-40254: Fix exposed seg fault in many loops specialization heuristic
in IP Cloning.

After inferPtrType() was changed to apply to typed pointers by getting the
actual pointer element type in the IR, a bug in the many loops specialization
heuristic code was exposed. DL.getTypeStoreSize(Ty), where Ty is the pointer
element type of an argument may not be defined when the argument is not marked
as dereferenceable. So we must check for !Arg.getDereferenceableBytes() before
we check for Arg.getDereferenceableBytes() < DL.getTypeStoreSize(Ty).

* Change requested by Chris