[VPlan][inscan] Change ReductionFinal operand to be the carry over value (#6944)
and adjust the carry-over computation for exclusive scan carry-over.
Previously the operand of reduction-final-inscan was load from reduction
memory, however, in masked mode loop it is not guaranteed that last lane
was unmasked. That means that when reduction-final-inscan extracts last
vector lane it may not have the reduction value.
It is easier to just pass the scalar carry-over value to
reduction-final-inscan, which is the value of the reduction. Then code
generation for reduction-final-inscan would be just returning the scalar
operand.

Exclusive scan carry-over needs to have the last input lane added to the
last lane of running reduction, so the next iteration would start from
it. It is not yet possible to add a test for this as exclusive inscan
loop rotation is not implemented yet.