[VPlan][HIR][CompressExpand] Improvement to compress expand idiom recognition to handle cases outside if (#7300)
* [VPlan][HIR][CompressExpand] Improvement to compress expand idiom recognition to handle cases outside if

HIR framework may fail to build HLIfs in certain cases and if this happens we
may fail to recognize compress/expand idiom as we were expecting the increment
and load to be inside an HLIf. This causes us to fail recognizing the idiom in
several hot loops in the lammps benchmark. To fix this, we now use
postdominates(inst, loop_first_child) to check that inst is conditional. We
also needed to remove the bailout in HIR vector code generation for non-search
loops when it saw labels and gotos. This is only done for loops with compress/
expand related instruction to avoid regressions else where.

This fixes CMPLRLLVM-35105 and should cover the 1% performance regression in
CMPLRLLVM-40130.

* Address review comment and undo an unnecessary code movement