[VPLoopAnalysis] Fix for type mismatching on induction instructions (#7320)
* [VPLoopAnalysis] Fix for type mismatching on induction instructions

This patch fixes an issue where an ExternalDef is used as a stride
for an induction and it's type is not the same as the start value
used for the induction. When VPlan generates VPInstructions for
induction updates and VPOCodeGen generates vector code for the
induction updates, there were type mismatches and verification
was failing. In this patch, the stride value is converted during
the creation of VPInductionInit and VPInductionInitStep instructions
so that the type is corrected initially and subsequent type conversions
at each point in VPlan are not needed.

* Address code review comments