[VPlan][Evaliator+CM] Reduce acceptable gain for vectorized remainder, add ZTT cost. (#7262)
* [VPlan][Evaliator+CM] Reduce acceptable gain for vectorized remainder, add ZTT cost.

Added a heuristic in evaluator:
The gain of vectorized remainder, i.e. (vector_cost-scalar_cost), is compared
with the scaled scalar cost and if it's less than the scaled value then the
scalar remainder is preferred. The scales are separate for masked and unmasked
remainders. When the remainder is enforced the pure gains of masked and
unmasked remainders are compared.

Added ZTT cost calculation in the cost model (as a separate interface), it's added
for unmasked vectorized remainder cost as we will insert one more TC check before
scalar remainder execution.

* Tests correction to reflect remainder vectorization enabling and optreport changes.

* Slight refactoring. Changed masked threshold.

* Adding pumping heuristic: in masked mode loop add pumping overhead to each
masked instruction.

* Changed thresholds for remainder gains.

* Responding review comments.

* One more review comment.

* More corrections on review.

* Increasing remainders threshold and responding last review comments