[SLP][NFC] Replace std::pair<ValueList, struct EdgeInfo> with dedicated struct. (#7368)
This minor refactoring is intended to address a build problem with gcc 7.5.0.
Speifically attempt to declare SmallVector of the pairs as a MultiNode class data
caused an error "constructor required before non-static data member for
'llvm::slpvectorizer::BoUpSLP::EdgeInfo::EdgeIdx'"
It looks like there is a gcc bug hit with that version.
If replace
     EdgeInfo() = default;
with
     EdgeInfo() {}
the problem goes away. By the way, gcc 8.5.0 successfully compiled the code.
As a workaround the above change would look ugly. Instead, OperandInfo structure
was introduced to convey that same information and interface was unified for
both reorderable and non-reorderable operands.