[OpenMP] Fix assert in codegen with depend iterators (#7396)
This is an additional fix for depend iterators. These variables are not
used directly to generate code and need to be ignored so they are not
emitted in implicit clauses. The previous change for this
(482d1f9e32695) registered them then checked when emitting a clause.

That doesn't go far enough since references are propagated to the outer
directives. This change instead prevents them from being added to the
references and therefore no implicit clauses are generated and they do
not make it to the outer directives.

Found in the OpenMP example tests:
  completed_openmp_c_cpp/tasking/positive/task_dep.11.c

Relates to: CMPLRLLVM-39225