[SYCL] Fix deprecation warning for headers (#6808) (#7486)
Using pragmas to emit warnings didn't work because SYCL headers are
considered to be system headers and any warnings in them are suppressed.

Use "#warning" instead. Unfortunately, MSVC doesn't support it (although
it's part of C23/C++23 and they'll have to add support eventually), so
we need some #if guards. Also, #warning cannot be put inside a macro
definition, thus we have to have some code duplication. Luckily, entire
headers deprecations aren't as often and we can be a little bit verbose.