[VPlan] Add idiom to search loop recognition for general value comparison (#7294)
* [VPlan] Add idiom to search loop recognition for general value comparison

Reference: https://jira.devtools.intel.com/browse/CMPLRLLVM-25459

The above JIRA reported a large performance difference between mainline
and xmain on the oneDAL benchmark because xmain doesn't recognize some
forms of search loops that mainline does.  The specific case we're looking
for is of the form:

  + DO i1 = 0, UB, 1 <DO_MULTI_EXIT_LOOP>
  |   if ((%a)[i1] <cmpop> %b)
  |   {
  |      %index = %i1;
  |      goto index_found;
  |   }
  + END LOOP

where (%a)[i1] can be executed speculatively, <cmpop> is any reasonable
comparison predicate, and %b is a loop-invariant expression.  As with
the PtrEq idiom, this "value comparison" idiom requires peeling to ensure
we don't read beyond available memory.

Alloy testing of this patch exposed a bug in HLLoop::generatePeelLoop()
where an expression created ahead of the peel loop was incorrectly marked
as linear.  The patch to that function included here was suggested by Satish
Guggilla and solves the problems I had observed.