[Paropt] Capture `schedule` chunk for passing it into the outlined region. (#7507)
JIRA: CMPLRLLVM-37507

The chunk value in `dist_schedule` clause was already getting captured before
passing it into the outlined function, but it was not happening for
the chunk of `schedule` clause.

We were previously using a hacky approach to sink the load instruction used to
compute the chunk into the region, but that is not correct for various reasons.

For example, it's possible that the source pointer from which the chunk value
is being loaded, gets modified inside the region, making hoisting the load
into the region illegal, since the chunk value should be the value the var
had before the region, and not a load from it after each chunk is done. e.g.

Src:
  #pragma omp parallel for schedule(static, var) shared(var)
  ... {var = 11 }

IR:
  %chunk.val = load i32, ptr %var
  ... [ DIR.OMP.PARALLEL.LOOP ... SCHEDULE(%chunk.val) SHARED(%var)
  store i32 11, %var ;

Another example is when the chunk's source pointer is captured on a
PRIVATE clause. The var won't retain the incoming value inside the region. e.g.

Src:
  #pragma omp parallel for schedule(static, var) private(var)

IR:
  %chunk.val = load i32, ptr %var
  ... [ DIR.OMP.PARALLEL.LOOP ... SCHEDULE(%chunk.val) PRIVATE(%var)

Now we capture the original value, and pass it into the region.