[OpenMP] Fix coverity issue with possible uninitialized LowerBound (NFC) (#7527)
The static analysis cannot tell that parameters would not be passed
that caused LowerBound to be uninitialized. Move the use of LowerBound
up so this is clear.

This code will need a complete cleanup when we switch to opaque pointers
permanently, so no need to spend too much time on this now.