[VPlan] Allow large single-iteration loops to be completely unrolled (#7553)
* [VPlan] Allow large single-iteration loops to be completely unrolled

After vectorization, a loop may only execute for a single iteration.
Currently we disable complete unrolling of a vectorized loop if the
instruction count in the loop body is larger than a threshold value,
even for single-iteration loops.  This patch relaxes the restriction
when we know there's a single iteration.