[OpaquePointer] Mangle the name of __spirv_ocl_printf correctly in Paropt. (#7528)
The SPIR-V translator relies on Itanium demangling to get the proper types of
function pointer arguments in opaque pointer mode. However, the name mangling
that paropt was using was implied to be (char AS2*, int) instead of the correct
(char AS2*, ...), and the mismatch in name mangling caused the translator to
back off of inferring names correctly. This fixes it to use the correct name,
and is one of the pieces needed for CMPLRLLVM-39793.