Generate (GEP, 0, 0) for call arguments (#7617)
[Opaque Pointers] (CPU2017ref/523) (CMPLRLLVM-40651)
This is needed to enable MemInitTrimDown & SOAToAOS to detect functionalities of member functions. Also needed to improve safety analysis.
(GEP, 0, 0) when type of argument of function is passed as subobject of the type to another call.
(GEP, 0, 0) when memory allocation pointer is passed as sub-object of the type to another call.