cmplrllvm-33925: Fix mate data when omp simd and unroll are both used. (#7650)
Clang generate hint to tell the BE vectorization is happen
before the unroll.

!8 = distinct !{!8, !9, !10}
!9 = !{!"llvm.loop.vectorize.enable", i1 true}
!10 = !{!"llvm.loop.vectorize.followup_all", !11}
!11 = distinct !{!11, !12, !13}
!12 = !{!"llvm.loop.isvectorized"}
!13 = !{!"llvm.loop.unroll.count", i32 4}

Our intel BE can not handle that.  But asked to remove the hint.

One way to do so:
Add new field IntelSimdVectorizeEnable in LoopAttributes, and
setIntelSimdVectorizeEnable. In CGStmtOpenMP.cpp, call
setIntelSimdVectorizeEnable instead setVectorizeEnable.

That allow us to skip community code of generating hint for BE.