[SYCL] Redesign pointer handling for OpenCL kernel generation (#7652)
* [SYCL] Redesign pointer handling for OpenCL kernel generation (#6728)

Requirement - Do not decompose types with pointers when generating
OpenCL kernel arguments.

This PR adds logic to stop decomposing trivial types containing
pointers. For every SYCL kernel argument which is a record type
containing a pointer (or has a field or a base class with a pointer), we
generate a new record type with all pointers in __global address space.
This compiler generated type is the openCL kernel argument. In the
kernel body, we initialize the local clone via memcpy.

Limitations:

1. Array of pointers or array of types with pointers are still
decomposed to it's elements.
2. Due to current implementation restrictions, types which are not
default constructible, continue to trigger decomposition if they contain
pointers.

Both limitations above will hopefully be fixed in follow-up PRs.

Signed-off-by: Elizabeth Andrews <elizabeth.andrews@intel.com>

* [NFC] Fix test (#6895)

size_t is different on windows and linux.

Signed-off-by: Elizabeth Andrews <elizabeth.andrews@intel.com>

* [SYCL] Do not decompose non-trivial classes with pointers  (#6886)

Instead, the following code is being generated:
```
void ocl_kernel(__generated_type GT) {
  Kernel KernelObjClone { *(reinterpret_cast<UsersType*>(&GT)) };
}
```

* [SYCL] Do not decompose non-trivial bases with pointers (#6964)

* Remove redundant test (#7532)

The case tested in this file is handled in SemaSYCL/inheritance.cpp.

Signed-off-by: Elizabeth Andrews <elizabeth.andrews@intel.com>
Co-authored-by: Mariya Podchishchaeva <mariya.podchishchaeva@intel.com>