[VPlan] Restrict complete unrolling of vectorized loops to loop depth > 1 (#7734)
PR 7553 allows large single-iteration loops to be completely
unrolled afterwards.  This had an adverse effect on cpu2017/525.x265.
The resulting very large basic block went through more optimizations,
removing some stores that had previously anchored useful SLP
vectorization.  There doesn't appear to be an easy way to deal with
the fallout.  Instead, this patch restricts the previous change to
only be allowed inside an outer loop.  This recovers the performance
for x265.

We might be able to relax this restriction in the future if SLP is
integrated with the loop optimizer.

Reference: https://jira.devtools.intel.com/browse/CMPLRLLVM-40696