[Driver][SYCL] Improve way sycld is pulled in for Linux based Windows… (#6974) (#7738)
… driver

Update the way sycld is pulled in by matching if msvcrtd is brought in
via the command line. Cmake uses '-Xclang --dependent-lib=msvcrtd' when
there is the desire to link in with the debug libraries. Check for this
and use an equivalent '--dependent-lib=sycld' to pull in the SYCL
runtime library.

This is only done for the Linux based drivers on Windows. The MSVC based
driver behavior already pulls in the proper libraries given /MDd.

Addresses: CMPLRLLVM-40401
Cherry-pick of: https://github.com/intel/llvm/commit/ebf6c5978cca731af907913961231769efbab007