review/haonanya/haonan/cherry pick (#7801)
* [SYCL] Disable  the LIT test 'debug-info-file-checksum.cpp' for Windows. (#6938)

The LIT test is using append-file. This command seems to have some
escape character issue on Windows.
Talked off-line to @pvchupin and agreed that for now the test case can
be disabled for windows.

Signed-off-by: Zahira Ammarguellat <zahira.ammarguellat@intel.com>

* [SYCL] Change "static constexpr int" -> "constexpr int" (#6984)

in sycl/source/detail/image_impl.hpp. clang-cl and MSVC produce
different exports for this. However, the static here doesn't bring any
functional value, so just remove it as a quick solution to align the
behavior between two toolchains that could be used to compile the
project.

The change is expected to be NFC for the MSVC toolchain used in our CI.

* [SYCL] Ignore some symbols during ABI check (#6993)

MSVC and clang-cl have minor differences in how the exported symbols are
created under some circumstances. We hit some of them in SYCL RT but we
believe they don't affect customers and so could be excluded from the
ABI check so that it would pass when the library is compiled with either
toolchain.

Co-authored-by: Steffen Larsen <steffen.larsen@intel.com>

Signed-off-by: Zahira Ammarguellat <zahira.ammarguellat@intel.com>
Co-authored-by: Zahira Ammarguellat <zahira.ammarguellat@intel.com>
Co-authored-by: aelovikov-intel <andrei.elovikov@intel.com>
Co-authored-by: Steffen Larsen <steffen.larsen@intel.com>