review/steffenlarsen/sycl host device removal (#7492)
* [SYCL] Fix unittests on MSVC (#6805)

This commit makes two changes:
* Fixes a cast of incompatible size in the PI mock plugin that caused a
warning on MSVC.
* Changes the definition of the captureless lambda function in the
RedefineAPI PiMock unittest to avoid MSVC considering it equal to the
function defined by the PI mock plugin.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>

* [SYCL] Make host device inaccessible through SYCL API (#6685)

This commit removes the host device from the device list and as such the
host device will no longer be available in user code. The following
changes are a result of this:
* Device filters using 'host' as either backend or device type will
cause a warning at runtime. Since there is no host device selectable for
these filters, the resulting device list will not contain a host
   device.
* is_host() on SYCL objects has been deprecated. Any use of them
internally on a host device should cause an assertion to fail.
* host_selector deprecation message has been changed to better reflect
that there is no alternative.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>

* [SYCL] Remove device-dependent tests from in-tree LIT (#6829)

This commit removes the currently failing in-tree LIT tests that were
previously dependent on a device being present. The removed tests are
either moved to the test-suite or were already there.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>

* [SYCL][NFC] Fix unused argument warning in host_selector (#6831)

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>

* [SYCL] Make the mock plugin report cl_khr_il_program as supported (#6823)

To make build-log tests not depend on the binary type, this commit makes
the mock plugin report that the cl_khr_il_program is supported. Together
with adding PI_DEVICE_INFO_NAME to mock_piDeviceGetInfo, the
corresponding redefinition from the build-log tests can be removed.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>

* [SYCL][NFC] Reintroduce missing Windows symbols (#6830)

https://github.com/intel/llvm/pull/6685 unintentionally caused the
windows library to lose symbols for handler::verifyKernelInvoc and
handler::is_host. This patch introduces these as unused members to avoid
the patch being ABI-breaking.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>

* [SYCL] Clear plugins after unload (#6846)

Unloading plugins was recently separated from the cleanup phase to allow
the mock plugin ensure that the mock plugin is the lone plugin. However,
after unloading the plugins vector is not currenly cleared. This commit
clears the plugins vector after plugins have been unloaded.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>

* [SYCL] Fix bugs in PI Mock (#6852)

* [SYCL] Fix constexpr string size in mock plugin (#6866)

https://github.com/intel/llvm/pull/6852 incidentally used a
non-constexpr solution for determining the size of a string, which
causes failures in select build configurations. This commit amends this
by using sizeof on a static array of characters instead of std::strlen.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>

* Remove rogue merge mistake in assert.cpp

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>
Co-authored-by: Romanov Vlad <vlad.romanov@intel.com>