[VPlan][HIR] Ensure that last value computed for FP IV is written back to original temp. (#7619)
Currently during HIR vectorizer codegen for induction-final the
last value is written back to arbitrary l-val. This is okay for
in-memory inductions, but incorrect for cases where the linear/FP
IV temp is liveout. This patch fixes the bug by ensuring that
last value is written back to original temp.

JIRA: CMPLRLLVM-40462