[Driver][SYCL] Improve way sycld is pulled in for Linux based Windows… (#6974) (#7738) (#7790)
… driver

Update the way sycld is pulled in by matching if msvcrtd is brought in
via the command line. Cmake uses '-Xclang --dependent-lib=msvcrtd' when
there is the desire to link in with the debug libraries. Check for this
and use an equivalent '--dependent-lib=sycld' to pull in the SYCL
runtime library.

This is only done for the Linux based drivers on Windows. The MSVC based
driver behavior already pulls in the proper libraries given /MDd.

Addresses: CMPLRLLVM-40401
Cherry-pick of: https://github.com/intel-restricted/applications.compilers.llvm-project/commit/81ea81a08fe69ff069f9aa652ed453de4702d285