[NFC][VPlan] Test showing illegal transformation by GVN (#7815)
* [NFC][VPlan] Test showing illegal transformation by GVN

This test shows an illegal transformation that is made by GVN
with respect to vectorizing loops with calls to vector functions
with vector of ptr args. Once the original function arg is widened,
any tranformations that rely on AA/ModRef and/or MemorySSA will
apply incorrect transformations. This happens because the type is
no longer ptr.

* Fix bug in the test. Stores using undef were showing incorrect
behavior due to the source code being used.