[SYCL] Optimize operator[] method of host accessor (#6853) (#7778)
* [SYCL] Optimize operator[] method of host accessor (#6853)

Compiler cannot optimize out calls to getPtr(), getOffset() and
getMemoryRange(), that were used in accessor::operator[], because they
may have side effects(e.g. printf). The patch adds a member in the
accessor which points to the actual data stored in the accessor_impl.
This pointer is initialized once during construction and used instead of
functions mentioned above.

* [SYCL] Fix post commit fails (#6875)

1. Updated checks to be more be more robust and work when clang discards
names of values in the LLVM IR.
2. Added a new symbol to the windows sybmols test.

* [SYCL][NFC] Add a missing symbol (#6885)

While the copy assignment operator of AccessorImplHost is not used, it
was auto generated by the compiler until recent changes and being
checked in the ABI tests.
The safest way to fix this is to just manually define it.

Also removed 2 redundant lines in host_acc_opt test.

* [SYCL] Relax checks in host_acc_opt LIT test (#7027)

It turns out to be non-robust to expect compiler to vectorize the code.
The code can be optmized in different ways. So relax the checks to
verify that there are no calls to accessor helper methods (getOffset and
getMemoryRange) in the loop. Absence of these calls should be enough to
optimize the code.