[Paropt] Add ability to specify RegisterAllocMode in VPOParoptConfig. (#7736)
For PVC, kernels can be compiled in either double grf mode or single grf mode. Currently, it is possible for the user to specify this per-kernel in OpenCL and SYCL (with pending change in intel/llvm by me). We also want to add support in OpenMP.

The approach is to extend VPOParoptConfig to have a RegisterAllocMode option.

`RegisterAllocMode` = `large` means ask IGC to use double GRF mode for this kernel.
`RegisterAllocMode` = `small` means ask IGC to use small double GRF mode for this kernel.
`RegisterAllocMode` = `auto` means ask IGC to use "auto" GRF mode, i.e. decide based on heuristics.
`RegisterAllocMode` = `default` means don't ask IGC anything specific per kernel, and leave it up to the other IGC compilation options or IGC's defaults to itself to determine what to do.

We add an argument to the metadata of either true or false to represent the YAML setting.

This change implements that design by extending the YAML format, adding code to get the status of the YAML option, and then set the `RegisterAllocMode` metadata option on kernels based on the yaml value. llvm-spirv knows what to do with this metadata (https://github.com/intel-innersource/applications.compilers.source.llvm-spirv-translator/pull/234) 

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>