[SLP] scheduler fixes for MultiNode support (#7929)
Basically two issues were found: (1) def-use lists to find dependencies
must be conducted some way to the scheduler in order to take multi-node
reordering into account, and (2) assumption that if an instruction is
scheduled as a single instruction (i.e. not in a bundle), then its
operands are not reordered, does not hold.