Add Intel Customization flag/docs for floating point comparison addition (#7996)
Fast-math is enabled by default in ICX. That entails some optimizations to
apply. Some of these is the removal of explicit comparisons to NaN and INF  
in the code. That leads to incorrect users. In order to help users
understand why their code is getting incorrect results (in comparison to
ICC), icx has a diagnostic that warns if the code contains explicit INF/NaN
comparisons either by value (NaN/INF) or by function (isnan/isinf). The
patch updated the customization following the model our customization and
changed the comment for the warning as it wasn't quite correct.
See https://www.gnu.org/software/libc/manual/html_node/Infinity-and-NaN.html
for details.