[CMPLRLLVM-40986][DTRANS] Handle Inalloca parameters. (#7962)
* [CMPLRLLVM-40986][DTRANS] Handle Inalloca parameters.

As required on Win32 for CMPLRLLVM-40986, this patch adds support for
inalloca parameters. A previous 'best guess' at how it works was
incorrect, since an inalloca parameter is actually a literal type filled
with all of the parameters, passed as a pointer.  The previous attempt
assumed that it was just the type itself, not realizing that inalloca
means that a list of parameters are passed in one llvm parameter.

* Fix bad #ifdef