[NFC] CMPLRLLVM-41169: Inlining report improvements: Factor call/function removal (#8084)
* [NFC] CMPLRLLVM-41169: Inlining report improvements: Factor call/function removal

This is part 1 of a series of inlining report fixes and improvements. This
NFC factors out the removeCallSiteReference() and removeFunctionReference()
in the metadata inlining report so they can be called in tandem with the
member functions of the same name for the classical inlining report.

I have removed the second argument which explictly passed the metadata
because it was not needed. In the inliner, I added explicit calls to
removeFunctionReference() before the function was removed from the module
in the old pass manager. This was needed because it severed the Function
from the module metadata. It will no longer be needed when we no longer
support the old pass manager.

* Fix minor errors. Rename function for consistency.

* Make comments consistent

* Fix for Prasad's comment