[OpenMP][Offload] Add support for dedicated zero-initialized memory pool (#7953)
Dedicated counter pool will be used when its map type is private and
zero_init (tentative). This pool can only handle counter-like data that
are less than 64B, and we do not expect a large number of live
allocations. Zero-initialization only happens when there is a new
allocation from L0, and device code is responsible for setting it back
to zero before releasing the memory.
Also added support for handling another new map type that forces use
of host memory.