[OpenCL] Fix work_group_broadcast produces wrong results w/o explicit barrier (#7805)
### Summary

This is the patch for [CMPLRLLVM-40078](https://jira.devtools.intel.com/browse/CMPLRLLVM-40078), with the help of #7777.
This patch modified GroupBuiltinPass to insert finalize_work_group_broadcast after work_group_broadcast, and some related tests.

### Why this happened

The bug can be simplified by this pseudo-code:
```
work_item = 1;
acc[] = {0, 0, ...}
loop {
    ret = group_broadcast(local_id, work_item);
    if (ret) break;
    acc[local_id] += 1
    if (local_id == 0) work_item = 0;
}
```
On CPU side, the work goup functions are converted into per-work-item function, which need accumulates the result in work goup Loop(wg-loop).

After BarrierPass, it is converted to 
```
v[WG_SIZE] = {1, 1, 1, …}
wg_loop_1 {
    ret = group_broadcast(local_id, v[local_id], &AllocaWGResult);
}
wg_loop_2 {
    if (!ret) break;
    acc[local_id] += 1
    if (local_id == 0) v[local_id] = 0; 
    ret = group_broadcast(local_id, v[local_id], &AllocaWGResult);
}
```
The ret is set to 1 after wg_loop_1.
In wg_loop_2, when local_id = 0 the ret is set to 0, then local_id = 1, and if condition is satisfied, thus break the loop.
Therefore acc[local_id] += 1 is not executed when local_id = 1(i,e, the second iteration).

That is, the `ret` used in `if condition` is the intermediate return value of per-wi group_broadcast.

### How to fix it

In GroupBuiltinPass.cpp, there was a process that for uniform & vecterized WG functions(except broadcast), a finalize-helper function will be inserted after the group_barrier of wg functions, which is used for summarizing the result of the vectorized wg function and converted into the constant extended value after wg_loop. That is, the finalize-helper function is used for synchronizing the final result of wg functions, but the GroupBuiltinPass wrongly assumed that finalized-helper function is required for all vectorized wg functions except broadcast.

That's it.