[VPlan] Fix for CMPLRLLVM-38438 (#7885) (#8027)
* [NFCI][VPlan] Fix for CMPLRLLVM-38438

There is a 4% performance regression in namd that is related to
loops with a high number of gathers and a chain of 3 vconflict
idioms. More investigation is necessary to determine why this
regression is happening, but this JIRA is on the MFL so a workaround
is provided here. The change is marked as NFCI because the fix is
highly targeted to these namd loops and therefore shouldn't trigger
a regression elsewhere. Creating a test would take quite a bit of
time and since this a temporary solution, the effort didn't seem
worth it.

* Add internal switches to control vectorization of scenarios including
high number of gathers and chain of vconflict. Also added test.

* Add positive lit test