[VPlan][inscan] Perform exchange of input and scan phases for exclusive scan loops (#7040)
Exclusive scan loop reduction can be transformed into inclusive reduction
with modified running reduction calculation. The key transformation is
to exchange the input and the scan phases of the inscan loop.
For exclusive scan loop scan phase precedes the separating pragma,
the input scan follows the separating pragma.

For example,

int x = start;​
#pragma omp simd reduction(inscan, +: x)​
for (int i = 0; i < N; ++i) {​
  out[i] = x;     // scan phase, before exclusive separating​ pragma​
#pragma omp scan exclusive(x)​
  x += in[i];     // input phase, after​ exclusive pragma​
}
At the point of writing x to out[i] we must write to ‘out’ pointer, but the
input to the running reduction is only yet to be computed.

To eliminate such dependency, flip the input and scan phases around the
directive and make this essentially an inclusive scan with shifted running
reduction computation.