review/erichkeane/cmplrllvm 41283 (#8252)
* [DTRANS] Ensure virtual-thunk-calls on MSVC are properly marked

The call generated by a virtual-function-thunk on MSVC was not properly
having its call-inst annotated with the DTrans metadata.  This patch
adds it.

As reported in: CMPLRLLVM-41283 (example b.cpp)

* [DTRANS] Ensure that we correct emit the microsoft VTable

As reported in CMPLRLLVM-41283, our Windows implementation of dtrans
metadata didn't emit the metadata for the vtable.  This patch fixes
that.

* Fixed clang-format issues