[NFC][VPlan][VectorUtils][TTI] Update interface for matching vector variants (#8166)
* [NFC][VPlan][VectorUtils][TTI] Update interface for matching vector variants

This patch provides the necessary provisioning to be able to do proper
matching for linear reference parameters. Namely, the addition of being
able to distinguish between linear reference arguments that are based on
private memory. E.g., to do proper matching, we must be able to
differentiate between things like 'foo(a[i])' vs. 'foo(i)'. For the
former case, we need to match with any available ref modifier variants.
For the latter case, we need to match with any available val/uval modifier
variants. Essentially, the difference being a linear pointer vs. a linear
pointee value.

* Address review comment