[HIR][Paropt] Don't crossover regions while finding region entry/exit directives for a given loop. (#8120)
We have a traversal algorithm in HIRFramework and VPOAnalysisUtils
that walks the CFG to identify the entry/exit directives that
correspond to given Loop. This patch fixes this algorithm to ensure
that we don't accidentally crossover from one region to another
(maybe siblings). Additionally we consolidate all users of this
algorithm to use common interface now defined in VPOAnalysisUtils.

JIRA: CMPLRLLVM-40970, CMPLRLLVM-40971