CMPLRLLVM-41169: Inline report improvements: Mainly for scalar opt (#8310)
* CMPLRLLVM-41169: Inline report improvements: Mainly for scalar opt

The xmain inlining reports have the additional complication that inlining is
an SCC pass and scalar optimizations are run between the inlining of SCCs
in the call graph. Often, an optimization will create a new function based
on a previous one and/or create new calls sites from old ones. If these are
not noted as connected old and new pairs, the inlining report will see them
as newly created and deleted functions and callsites.

This PR uses standard inlining report APIs to indicate old and new functions
and call sites to the inlining report. Most of these changes are made to
Scalar Opt. There are also some minor changes that were made to the inlining
reports to accomodate them. I have also indicated places in Scalar Opt where
calls are deleted through dead code elimination.

The changea are not exhaustive, but cover the cases in telecom/convendata00.
They are a good starting place. Additional instrumentations can be done as
time permits.

Apologies for clang formating the files. This is now stringly encouraged
now that an automatic tool is flagging clang-format violations.

* Calng format changes

* Remove stray 'slop' file

* Add INTEL_CUSTOMIZATION

* Revert global changes with clang-format to community files.

* Revert global changes from clang-format on community files

* Apply clang-format diff

* Fix for Prasad's comment

* Fixes requested by Chris