[SYCL] Native event for default-ctored sycl::event has to be in COMPLETE state (#7148) (#8290)
Per SYCL 2020 for event():

   > Constructs an event that is immediately ready. The event has no
   > dependencies and no associated commands. Waiting on this event will
   > return immediately and querying its status will return
   > info::event_command_status::complete.

Modify piEventCreate to create an event in such a state.

There is a more general problem that isn't addressed here:

  auto e = q.submit(... h.host_task(...) ..)

This event would be a host one and we assert that no get_native could be
called on it (see existing sycl::detail::getImplBackend). If we will
ever want to support such scenario we'd need to implement some tracking
of host/backed events in the SYCL RT and keep updating the latter
whenever the host one changes the state.

Alternatively, SYCL spec could be updated to prohibit such scenario or
specify that such event has no native counterpart.