CMPLRLLVM-39714: Threshold map should increment also when a switch/case is unswitched (#8154)
This change set updates the threshold map in HIROptPredicate
to handle the switch/case instructions. The map will increase
the number of copies per loop by the number of cases that where
unswitched if there aren't any If instructions inside the loop.
Else, the threshold won't increase and will be handled by the
transformation tha unswitches the If conditions.