CMPLRLLVM-39845: Restore IPcloning for 538.imagick after pulldown (#8431)
This restores the dead arg elimination pass that is needed to get the IR
in the right form for IP cloning for 538.imagick. This pass was moved by
the community in the last pulldown. I looked into changing IP cloning, itself,
but the control flow after this change is compilcated by the insertion of
select in awkward places that obscure the control flow analysis in the
IP cloning. So putting this dead argument elimination in the old place for
DTrans only is the simplest fix.