[NFC] Compile time improvements for better top sort number assignment (#8355)
There is a test case in CMPLRLLVM-37390 where we were hanging in
opt-predicate on a big loopnest. Previously, I had suppressed formation
of the loop as a workaround.

The opt command line takes at least 54 minutes for the big loopnest. It
looks like we do a lot of unswitching. The original loopnest is a little
over 1300 lines of HIR. After predicate-opt it increases to over 350,000
lines. I think part of the problem is that we don't have threshold of
unswitching switches. Axel is working on this part in PR #8154 so that
should help.

I was looking into top sort number computation because according to perf
tool this part of the code was taking 99% of the time.

This PR has the following changes-

1) We use half the default spacing for the slow path of top sort number
recomputation when the number of nodes don't fit in available range.
Previously, we were using a spacing of 1 which increases the probability
of recomputation due to further nodes insertions. This alone bring the
compile time of the opt command to less than 9 minutes.

2) Added a command line option for experimenting with top sort num spacing.

3) Added a check for prev/next nodes in replace*() utilities to give the
new nodes a bigger top sort num range which reduces the probability that
new nodes will not fit in available top sort range. This saves another 30
seconds of compile time.